requirements.txt


pycrypto
